Move properties common to both AD schemes to AddressDerivationSpec
