+++ lib/core/cardano-wallet-core.cabal
@@ -186,6 +186,7 @@ test-suite unit
       Cardano.Wallet.DBSpec
       Cardano.Wallet.DummyTarget.Primitive.Types
       Cardano.Wallet.NetworkSpec
+      Cardano.Wallet.Primitive.AddressDerivationSpec
       Cardano.Wallet.Primitive.AddressDerivation.SequentialSpec
       Cardano.Wallet.Primitive.AddressDiscovery.SequentialSpec
       Cardano.Wallet.Primitive.CoinSelection.LargestFirstSpec
